generate PDL that's correct upto semantics
Now we get the error:

temp.mlir:8:15: error: 'pdl.operation' op expected a bindable (i.e. `pdl.operation` or `pdl.rewrite`) user when defined in the matcher body of a `pdl.pattern`
      %zero = "pdl.operation"(){

This tells us that we do not generate the right PDL, which
is indeed true. But our PDL is at least well formed (we pass the low
level verifiers!)