mlir: prototype SSA environment and dialect definitions
I explore a system in which dialect operations are defined through an
automatically-generated inductive type, which is much easier to match
than the generic operation format (even with dependent constraints).

The elegance of this approach lies in the fact that such an inductive
type is in fact an event family, which means that MLIR programs can be
fully-automatically translated into ITres by emitting events for each
operation. The user can then define the semantics of the dialect as an
event handler. This has several benefits:

(1) It is much more user-friendly.
(2) The auto-translation from the dialect into ITrees can handle name
    queries to the SSA environment.
(3) It makes the composition of dialects a trivial matter by delegating
    it to the usual handlers' composition of monad transformers.

The changes for this experiment include:

* MLIRSemantics.Fitree: define event families as polymorphic (with type
  [Type → Type u] instead of [Type → Type]) to allow operations'
  semantics to quantify over types.
* MLIRSemantics.Toy.ToyDialect: show an idea of what auto-generated
  definitions could look like.

I also outlined a first approach to the SSA environment, which basically
consists of (1) we don't enforce SSA constraints and define a normal
imperative-style environment, (2) we evaluate rewrite patterns in this
environment, which works fine because there is no interference, (3) we
prove that when SSA form is used, the "imperative" semantics remain
valid even with other operations mixed in.

* MLIRSemantics.SSAEnv: prototype SSA environments.

Finally, some housekeeping:

* MLIRSemantics.Verifier: move verifier definitions to a new file.