Introduce mlir_string which can accept escaping
MLIR escapes strings within strings by hex-encoding the
offending characters. Eg: chr(0x22) = ", so MLIR encode
the literal:

> "foo"bar"baz"

as:

> "foo\22bar\22baz"

But Lean does not like the `\2` in a string, which it
attempts to interpret as an illegal escape character.

To eliminate these limits, we drop into the lowest level of the parser,
and define a custom `mlir_string` type that allows such escaping.