util: fix the simp_itree/dsimp_itree tactics
I suspect the errors were due to commas missing between the main
lemmas, resulting in an even number of child nodes to [simp] (which is
impossible normally). Although that would suggest that half of the
registered @[simp_itree] lemmas would not apply, and I have never
observed any behavior of that kind.