semantics: improve UB, half give up on rewrite-style proofs
Improved UBE to have a single event and return any α, which is possible
without a universe bump if all constructors have it (probably because
Lean moves it to an index).

For some reason rewrite-style proofs trigger the bug in th2 but the
direct unfolding-style doesn't. I'd rather have the correct semantics
than a faster/more convenient proof, so I'll leave it there.