semantics: interpret UB last
This is the proper thing to do since in the future we want dialect
handlers to be able to emit UB. It's also helpful in proofs since we
don't have to sift SSA events through interpUB' at the first layer.

Amazingly, this commit is accompanied by a full refresh of the proofs
in ArithSemantics.lean, with only minor encounters of the WF induction
bug.

Ultimately simp_itree remains more efficient in development effort
than all the ops.*.sem theorems, which just re-state what we've already
written. It would be nice to get rid of these lemmas, either with
simp_itree, or with some conv + unfolding the ArithE handler.