Remove handle.
We drastically simplify the model by removing `handle`.
We shall next defunctionalize `Fitree`, hard-coding SSA
effects into our `Fitree`. This means that we will not
need to reason about `Vis` in general, but only in particular
about `Read, Write` (from `SSAEnvE`), plus `UB` from `UBE`.

We will retain a notion of extensibility of semantics, because
we will describe how to graft together different handlers for
operations. This will be the work of our `Semantics S` typeclass,
which will have instances for sums of dialects (`d1 + d2`).